audio_transcriber/requirements.txt
flask
yt-dlp  
google-cloud-speech>=2.0.0
python-dotenv
confluent-kafka
